# shared/openapi.yaml
openapi: 3.0.3
info:
  title: AlpRoutes API
  version: 0.1.0
  description: |
    Единый API для Android- и веб-клиента каталога альпинистских маршрутов.
    Этот файл — источник истины: клиентские модели генерируются из него.

    ## Общие правила

    * **Геометрия** — GeoJSON, EPSG:4326, порядок координат `[lon, lat]` или `[lon, lat, ele]`.
      `bbox` в запросах — `minLon,minLat,maxLon,maxLat`.
    * **Пагинация** — у любого списка. Курсорная: `cursor` + `limit` (1–100, по умолчанию 20),
      в ответе `nextCursor` (null — страниц больше нет). Курсор непрозрачный.
    * **Мультиязычность** — тексты каталога (`LocalizedText`) приходят целиком, со всеми
      переводами: `{"ru": "...", "en": "..."}`. Язык выбирает клиент
      (порядок: язык интерфейса → `ru` → любой имеющийся). Пользовательский контент
      (комментарии, отчёты) не переводится и несёт поле `language`.
    * **Отсутствующие данные** — `null` означает «нет данных». Клиент показывает «нет данных»,
      а не ноль и не прочерк, похожий на значение. Сервер никогда не подставляет оценку.
    * **Категории сложности** — всегда пара `{system, value}`, у маршрута может быть несколько,
      не больше одной на систему. Конвертации между системами в API нет и не будет.
    * **Офлайн-запись.** Всё, что пользователь создаёт (восхождения, комментарии, фото, треки),
      создаётся с `id`, сгенерированным на клиенте (UUID). Повтор того же запроса с тем же `id`
...
    * **Ошибки** — `application/problem+json` (RFC 9457).
...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /auth/oauth/{provider}:
    post:
      tags: [auth]
      operationId: loginWithOAuth
      summary: Вход через Google или Apple
      description: |
        Клиент получает ID token у провайдера и передаёт его сюда. Сервер проверяет подпись,
        находит или создаёт пользователя и выдаёт пару токенов.
      security: []
      parameters:
        - name: provider
          in: path
          required: true
          schema:
            type: string
            enum: [google, apple]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/OAuthLoginRequest' }
      responses:
        '200':
          description: Успешный вход
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TokenPair' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /auth/dev-login:
    post:
      tags: [auth]
      operationId: devLogin
      summary: Вход без OAuth (только локальная разработка)
      description: |
        Существует только в профиле `dev`; в остальных окружениях эндпоинт не зарегистрирован (`404`).
        Находит пользователя по email или создаёт его, назначает указанную роль.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DevLoginRequest' }
      responses:
        '200':
          description: Успешный вход
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TokenPair' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
  /auth/refresh:
    post:
      tags: [auth]
      operationId: refreshTokens
      summary: Обновить пару токенов
      description: |
        Ротация: старый refresh token становится недействительным. Повторное предъявление
        уже использованного токена отзывает всю цепочку (признак кражи) — клиент должен
        выполнить вход заново. Исключение — повтор в течение 30 секунд после ротации
        (два параллельных запроса одного клиента): `401` без отзыва цепочки.
        Клиент обязан выполнять обновление в один поток.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RefreshRequest' }
      responses:
        '200':
          description: Новая пара токенов
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TokenPair' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /auth/logout:
    post:
      tags: [auth]
      operationId: logout
      summary: Выйти (отозвать refresh token)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RefreshRequest' }
      responses:
        '204': { description: Токен отозван }
...
  /routes:
    get:
      tags: [routes]
      operationId: listRoutes
      summary: Поиск маршрутов
      description: |
        Все фильтры комбинируются через И. Геофильтры выполняются в PostGIS.
        `bbox` ищет по всем геообъектам маршрута (маршрут попадает в окно, даже если
        в нём виден только подход), `near` — по точке маркера (`anchorPoint`).
        Маршруты без геометрии в гео-выдачу не попадают.
        Фильтр по категории работает только внутри одной системы:
        `gradeSystem=RU&gradeMin=4А&gradeMax=5Б`. Маршруты без категории в этой системе
        в выдачу не попадают — клиент должен сообщать об этом пользователю.
      security: [{}, { bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/Bbox'
        - name: near
          in: query
          description: Точка `lon,lat`; вместе с `radiusM`. Сортирует по расстоянию.
          schema: { type: string, example: '43.0,43.1' }
        - name: radiusM
          in: query
          schema: { type: integer, minimum: 1, maximum: 200000 }
        - name: areaId
          in: query
          schema: { type: string, format: uuid }
        - name: includeSubareas
          in: query
          description: Включать маршруты вложенных районов
          schema: { type: boolean, default: true }
        - $ref: '#/components/parameters/SearchQuery'
        - name: gradeSystem
          in: query
          schema: { $ref: '#/components/schemas/GradeSystemCode' }
        - name: gradeMin
          in: query
          description: Требует `gradeSystem`
          schema: { type: string }
        - name: gradeMax
          in: query
          description: Требует `gradeSystem`
          schema: { type: string }
        - name: routeType
          in: query
          schema: { $ref: '#/components/schemas/RouteType' }
        - name: sort
          in: query
          description: При `near` по умолчанию `distance`, при `q` — `relevance`, иначе `name`.
          schema:
            type: string
            enum: [name, grade, updated, distance, relevance]
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RouteSummaryPage' }
        '400': { $ref: '#/components/responses/BadRequest' }
    post:
      tags: [routes]
      operationId: createRoute
      summary: Предложить новый маршрут
      description: Создаёт маршрут в статусе `draft` и первую ревизию в статусе `pending`.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RouteCreate' }
      responses:
        '201':
          description: Создано, ожидает модерации
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RouteRevision' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409': { $ref: '#/components/responses/Conflict' }
  /routes/map-points:
    get:
      tags: [routes]
      operationId: listRouteMapPoints
      summary: Облегчённые точки маршрутов для карты
      description: Минимум полей для маркеров (`anchorPoint`). Маршруты без геометрии не возвращаются. Лимит страницы выше обычного.
      security: [{}, { bearerAuth: [] }]
      parameters:
        - name: bbox
          in: query
          required: true
          schema: { type: string, example: '42.9,43.0,43.3,43.2' }
        - $ref: '#/components/parameters/Cursor'
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 500, default: 200 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RouteMapPointPage' }
        '400': { $ref: '#/components/responses/BadRequest' }
  /routes/{routeId}:
    parameters:
      - $ref: '#/components/parameters/RouteId'
    get:
      tags: [routes]
      operationId: getRoute
      summary: Карточка маршрута (текущая опубликованная ревизия)
      security: [{}, { bearerAuth: [] }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RouteDetail' }
        '404': { $ref: '#/components/responses/NotFound' }
  /routes/by-slug/{slug}:
    get:
      tags: [routes]
      operationId: getRouteBySlug
      summary: Карточка маршрута по slug (для SSR-страниц)
      security: [{}, { bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/Slug'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RouteDetail' }
        '404': { $ref: '#/components/responses/NotFound' }
  /routes/{routeId}/revisions:
    parameters:
      - $ref: '#/components/parameters/RouteId'
    get:
      tags: [routes]
      operationId: listRouteRevisions
      summary: История правок маршрута
      security: [{}, { bearerAuth: [] }]
      parameters:
        - name: status
          in: query
          schema: { $ref: '#/components/schemas/ReviewStatus' }
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: OK, новые сверху
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RouteRevisionSummaryPage' }
        '404': { $ref: '#/components/responses/NotFound' }
    post:
      tags: [routes]
      operationId: proposeRouteRevision
      summary: Предложить правку маршрута
      description: |
        Правка — полный снимок маршрута, а не дифф. `baseRevisionId` — ревизия, от которой
        автор начал редактирование. Если с тех пор опубликована другая, ответ `409`
        с `currentRevisionId`: клиент показывает автору изменения и просит слить вручную.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RouteRevisionCreate' }
      responses:
        '201':
          description: Создано, ожидает модерации
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RouteRevision' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /route-revisions/{revisionId}:
    parameters:
      - $ref: '#/components/parameters/RevisionId'
    get:
      tags: [routes]
      operationId: getRouteRevision
      summary: Ревизия маршрута целиком
      security: [{}, { bearerAuth: [] }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RouteRevision' }
        '404': { $ref: '#/components/responses/NotFound' }
  /route-revisions/{revisionId}/approve:
    parameters:
      - $ref: '#/components/parameters/RevisionId'
    post:
      tags: [moderation]
      operationId: approveRouteRevision
      summary: Одобрить и опубликовать ревизию (модератор)
      description: |
        Ревизия становится текущей, маршрут — `published`. Если с момента создания правки
        опубликована другая ревизия (её `baseRevisionId` устарел), — `409` с `type = revision-stale`:
        одобрение затёрло бы чужие изменения. Такую правку нужно отклонить или пересоздать.
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReviewDecision' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RouteRevision' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /route-revisions/{revisionId}/reject:
    parameters:
      - $ref: '#/components/parameters/RevisionId'
    post:
      tags: [moderation]
      operationId: rejectRouteRevision
      summary: Отклонить ревизию (модератор)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReviewDecisionWithNote' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RouteRevision' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /route-revisions/{revisionId}/revert:
    parameters:
      - $ref: '#/components/parameters/RevisionId'
    post:
      tags: [moderation]
      operationId: revertToRouteRevision
      summary: Откатить маршрут к этой ревизии (модератор)
      description: |
        Создаёт НОВУЮ ревизию — копию указанной — и сразу публикует её.
        История не переписывается.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReviewDecisionWithNote' }
      responses:
        '201':
          description: Новая текущая ревизия
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RouteRevision' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }

  # ------------------------------------------------------------- uploads
...
  /moderation/route-revisions:
    get:
      tags: [moderation]
      operationId: listPendingRouteRevisions
      summary: Правки маршрутов на проверке (модератор)
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: OK, старые сверху
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RouteRevisionSummaryPage' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token из `TokenPair`. Живёт недолго, обновляется через `/auth/refresh`.
  parameters:
    Cursor:
      name: cursor
      in: query
      description: Непрозрачный курсор из `nextCursor` предыдущей страницы
      schema: { type: string }
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
    SearchQuery:
      name: q
      in: query
      description: Поиск по названию на всех языках (триграммы), минимум 2 символа
      schema: { type: string, minLength: 2, maxLength: 100 }
    Bbox:
      name: bbox
      in: query
      description: '`minLon,minLat,maxLon,maxLat` в EPSG:4326'
      schema: { type: string, example: '42.9,43.0,43.3,43.2' }
    Slug:
      name: slug
      in: path
      required: true
      schema: { $ref: '#/components/schemas/Slug' }
...
    RouteId:
      name: routeId
      in: path
      required: true
      schema: { type: string, format: uuid }
    RevisionId:
      name: revisionId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    BadRequest:
      description: Некорректный запрос
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Unauthorized:
      description: Нужна аутентификация или токен недействителен
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Forbidden:
      description: Недостаточно прав
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    NotFound:
      description: Не найдено
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Conflict:
      description: |
        Конфликт: занятый slug, чужой объект с тем же `id`, устаревшая базовая ревизия
        (тогда в `Problem.currentRevisionId` — актуальная) или недопустимый переход статуса.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
  schemas:
...
    Problem:
      type: object
      description: |
        RFC 9457. Значения `type`: `validation-error`, `unauthorized`, `forbidden`, `not-found`,
        `conflict`, `slug-taken`, `revision-conflict`, `revision-stale`, `invalid-state`,
        `track-not-contiguous`, `payload-too-large`, `internal-error`.
      required: [type, title, status]
      properties:
        type: { type: string, description: 'Машиночитаемый код, например `revision-conflict`' }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        errors:
          type: array
          items: { $ref: '#/components/schemas/FieldError' }
        currentRevisionId:
          type: string
          format: uuid
          description: Только для `revision-conflict`
...
    Slug:
      type: string
      pattern: '^[a-z0-9]+(-[a-z0-9]+)*$'
      maxLength: 120
      example: dykh-tau-mummery
...
    ReviewStatus:
      type: string
      enum: [pending, approved, rejected]
...
    OAuthLoginRequest:
      type: object
      required: [idToken]
      properties:
        idToken: { type: string }
        nonce: { type: string, description: 'Nonce, использованный при запросе токена (Apple)' }
        displayName:
          type: string
          maxLength: 100
          description: Apple отдаёт имя только при первом входе — клиент передаёт его сюда
    DevLoginRequest:
      type: object
      required: [email]
      properties:
        email: { type: string, format: email }
        displayName: { type: string, minLength: 1, maxLength: 100 }
        role: { $ref: '#/components/schemas/UserRole' }
    RefreshRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken: { type: string }
    TokenPair:
      type: object
      required: [accessToken, accessTokenExpiresAt, refreshToken, refreshTokenExpiresAt, user]
      properties:
        accessToken: { type: string }
        accessTokenExpiresAt: { type: string, format: date-time }
        refreshToken: { type: string }
        refreshTokenExpiresAt: { type: string, format: date-time }
        user: { $ref: '#/components/schemas/User' }

    # -------------------------------------------------------------- users
    UserRole:
      type: string
      enum: [user, moderator, admin]
...
    User:
      allOf:
        - $ref: '#/components/schemas/UserPublic'
        - type: object
          required: [locale, role]
          properties:
            email: { type: string, format: email, nullable: true }
            locale: { $ref: '#/components/schemas/LanguageCode' }
            role: { $ref: '#/components/schemas/UserRole' }
...
    GradeSystemCode:
      type: string
      pattern: '^[A-Z][A-Z_]*$'
      description: |
        Код системы из `GET /grade-systems`. Сейчас: `RU`, `IFAS`, `UIAA`, `FR_FREE`, `YDS`, `AID`, `WI`.
        Намеренно строка, а не enum: новые системы добавляются данными, и старые версии
        приложения не должны падать на незнакомом коде.
      example: RU
...
    ContentStatus:
      type: string
      enum: [draft, published, hidden]
...
    RouteType:
      type: string
      enum: [rock, snow_ice, combined]
      description: Скальный, снежно-ледовый, комбинированный
...
    RouteSummary:
      type: object
      required: [id, slug, status, name, area, grades, updatedAt]
      properties:
        id: { type: string, format: uuid }
        slug: { $ref: '#/components/schemas/Slug' }
        status: { $ref: '#/components/schemas/ContentStatus' }
        name: { $ref: '#/components/schemas/LocalizedText' }
        area: { $ref: '#/components/schemas/AreaRef' }
        grades:
          type: array
          items: { $ref: '#/components/schemas/Grade' }
        routeType: { $ref: '#/components/schemas/RouteType' }
        elevationGainM: { type: integer, nullable: true }
        lengthM: { type: integer, nullable: true }
        anchorPoint: { $ref: '#/components/schemas/AnchorPoint' }
        distanceM:
          type: number
          nullable: true
          description: Только при поиске с `near` — расстояние по geography
        coverPhotoUrl: { type: string, format: uri, nullable: true }
        updatedAt: { type: string, format: date-time }
...
    RouteDetail:
      allOf:
        - $ref: '#/components/schemas/RouteContent'
        - type: object
          required: [id, slug, status, area, areaPath, currentRevision, updatedAt]
          properties:
            id: { type: string, format: uuid }
            slug: { $ref: '#/components/schemas/Slug' }
            status: { $ref: '#/components/schemas/ContentStatus' }
            area: { $ref: '#/components/schemas/AreaRef' }
            areaPath:
              type: array
              description: Хлебные крошки от корня к району маршрута
              items: { $ref: '#/components/schemas/AreaRef' }
            currentRevision: { $ref: '#/components/schemas/RouteRevisionRef' }
            stats: { $ref: '#/components/schemas/RouteStats' }
            updatedAt: { type: string, format: date-time }
...
    RouteSummaryPage:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/RouteSummary' }
        nextCursor: { type: string, nullable: true }
...
    RouteMapPointPage:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/RouteMapPoint' }
        nextCursor: { type: string, nullable: true }
    RouteCreate:
      type: object
      required: [content]
      properties:
        slug:
          allOf:
            - $ref: '#/components/schemas/Slug'
          description: Необязательно — сервер сгенерирует транслитерацией названия
        content: { $ref: '#/components/schemas/RouteContent' }
        changeSummary: { type: string, maxLength: 500 }
        publish:
          type: boolean
          default: false
          description: |
            Только модератор: сразу одобрить и опубликовать (в журнал модерации пишется
            как самоодобрение). Для остальных `true` — `403`.
    RouteRevisionCreate:
      type: object
      required: [baseRevisionId, content]
      properties:
        baseRevisionId:
          type: string
          format: uuid
          description: Ревизия, от которой начато редактирование
        content: { $ref: '#/components/schemas/RouteContent' }
        changeSummary: { type: string, maxLength: 500 }
        publish:
          type: boolean
          default: false
          description: |
            Только модератор: сразу одобрить и опубликовать (в журнал модерации пишется
            как самоодобрение). Для остальных `true` — `403`.
    RouteRevisionRef:
      type: object
      required: [id, revisionNumber]
      properties:
        id: { type: string, format: uuid }
        revisionNumber: { type: integer }
        approvedAt: { type: string, format: date-time, nullable: true }
    RouteRevisionSummary:
      type: object
      required: [id, routeId, revisionNumber, author, createdAt, status, isCurrent]
      properties:
        id: { type: string, format: uuid }
        routeId: { type: string, format: uuid }
        revisionNumber: { type: integer }
        baseRevisionId: { type: string, format: uuid, nullable: true }
        revertedFromId: { type: string, format: uuid, nullable: true }
        author: { $ref: '#/components/schemas/UserPublic' }
        changeSummary: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
        status: { $ref: '#/components/schemas/ReviewStatus' }
        isCurrent: { type: boolean }
        reviewedBy: { $ref: '#/components/schemas/UserPublic' }
        reviewedAt: { type: string, format: date-time, nullable: true }
        reviewNote: { type: string, nullable: true }
    RouteRevision:
      allOf:
        - $ref: '#/components/schemas/RouteRevisionSummary'
        - type: object
          required: [content]
          properties:
            content: { $ref: '#/components/schemas/RouteContent' }
    RouteRevisionSummaryPage:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/RouteRevisionSummary' }
        nextCursor: { type: string, nullable: true }
    ReviewDecision:
      type: object
      properties:
        note: { type: string, maxLength: 2000 }
    ReviewDecisionWithNote:
      type: object
      required: [note]
      properties:
        note: { type: string, minLength: 1, maxLength: 2000 }

    # ------------------------------------------------------------ uploads
...